requirements.txt
langchain==0.1.11
openai==1.13.3
python-dotenv==1.0.1
langchain_openai==0.0.8
google-search-results==2.4.2
youtube_search==2.1.2